requirements.txt
fastapi
feedparser
httpx
pandas
passlib
python-dateutil
python-dotenv
python-jose
python-multipart
sqlalchemy==1.4.35
sqlmodel
typer[all]
uvicorn